Moved files into similar folders
