+++ src/components/MediaControlsBar/TrackProgress/ProgressBar/ProgressBar.css
