+++ src/components/MediaControlsBar/TrackProgress/TrackProgress.css
