+++ src/components/MediaControlsBar/TrackProgress/index.tsx
@@ -1,22 +1,24 @@
 import React, { Component, ReactNode } from "react";
-import { TIME_FORMAT } from "../../constants";
-import CurrentTime from "../CurrentTime";
-import Duration from "../Duration";
-import ProgressBar from "../ProgressBar";
+import { TIME_FORMAT } from "../../../constants";
+import CurrentTime from "./CurrentTime";
+import Duration from "./Duration";
+import ProgressBar from "./ProgressBar";
 import "./TrackProgress.css";
+import { I18nAriaLabels } from "../../../types";
 
 interface TrackProgressProps {
     audio: HTMLAudioElement;
     defaultCurrentTime?: ReactNode;
     defaultDuration?: ReactNode;
     timeFormat?: TIME_FORMAT;
+    i18nAriaLabels?: I18nAriaLabels;
 }
 
 class TrackProgress extends Component<TrackProgressProps> {
 
     progressRef = React.createRef<HTMLDivElement>();
     render() {
-        const { audio, timeFormat, defaultCurrentTime, defaultDuration } = this.props;
+        const { audio, timeFormat, defaultCurrentTime, defaultDuration, i18nAriaLabels } = this.props;
         if(!timeFormat) {
             return null;
         }
@@ -28,7 +30,7 @@ class TrackProgress extends Component<TrackProgressProps> {
                     timeFormat={timeFormat}
                     defaultCurrentTime={defaultCurrentTime}
                 />
-                <ProgressBar ref={this.progressRef} audio={audio} />
+                <ProgressBar ref={this.progressRef} audio={audio} i18nProgressBar={i18nAriaLabels?.progressControl} />
                 <Duration
                     className="media-controls-progress-time"
                     audio={audio}
