+++ src/components/MediaControlsBar/VolumeControls/VolumeBar/index.tsx
@@ -1,0 +1,12 @@
+import React, { Component } from "react";
+
+
+export default class VolumeBar extends Component {
+    render(): React.ReactNode {
+        return (
+            <div className="media-controls-volume-bar">
+
+            </div>
+        );
+    }
+}
