+++ src/types.ts
@@ -1,0 +1,15 @@
+export interface I18nAriaLabels {
+    player?: string;
+    progressControl?: string;
+    volumeControl?: string;
+    play?: string;
+    pause?: string;
+    rewind?: string;
+    forward?: string;
+    previous?: string;
+    next?: string;
+    loop?: string;
+    loopOff?: string;
+    volume?: string;
+    volumeMute?: string;
+}
